+++ changelog.md
@@ -1,8 +1,16 @@
+## version 1.3.0 [In Progress]
+1. Deprecated: `getdoc()`;
+2. Feature: support app['lessweb.apidoc.routes'];
+3. Feature: Support multiple positional-only parameters;  //TODO: need testing
+
+## version 1.2.0
+1. Feature: support typecasting Union and NewType;
+
 ## version 1.1.0
-1. Feature: add unittest (new tests/ dir)
-2. Feature: add mypy test
-4. Feature: add Github Action
-3. Feature: support typecasting Literal
+1. Feature: add unittest (new tests/ dir);
+2. Feature: add mypy test;
+4. Feature: add Github Action;
+3. Feature: support typecasting Literal;
 
 ## version 1.0.8